kernel: hid: revert uhid driver back to its initial state
For some reason the uhid driver was modified to expose an HID driver
with the same name 'generic-bluetooth' as the one provided by the
bluetooth subsystem. This causes the kernel failing to load the BT
HID one on startup which in turn causes BlueZ to fail setting up a
HID device from the userland side as the necessary protocol for
AF_BLUETOOTH BTPROTO_HID was never registered.

It's unsure who modified the uhid driver for whatever reason but
this needs to be reverted in order to get things properly working
with BlueZ.

Source of the driver taken from upstream v3.6 release as this is
the first the driver appears in and which looks most similar to
the version we have here in our 3.4 based kernel.

Change-Id: I82f711d419f46deb0fea256c796a3afc547f0be2
Signed-off-by: Simon Fels <simon.fels@canonical.com>