mediatek: config: introduce FORM_FACTOR variable
The FORM_FACTOR variable will be used for different aspects in the
system where we need to find out what kind of device we're running
on. One of the first things which will use this is the Bluetooth
subsystem in order to determine which class of device we have to
set for the device it's running on.

This also renames the file stored in repo to its proper name
android.conf as which it is install as later on.

Change-Id: I039a1c9f314f79ba02ec88e29747be812121d6d5
Signed-off-by: Simon Fels <simon.fels@canonical.com>